Move some libs to concerns on ApplicationRecord and stop pollute ActiveRecord::Base
