+++ app/models/concerns/dorsale/model_to_s.rb
@@ -1,0 +1,9 @@
+module Dorsale::ModelToS
+  def to_s
+    %w(name title label).map do |m|
+      return send(m) if respond_to?(m)
+    end
+
+    super
+  end
+end
