+++ app/pdfs/concerns/dorsale/alexandrie/prawn/render_with_attachments.rb
@@ -1,0 +1,13 @@
+module Dorsale::Alexandrie::Prawn::RenderWithAttachments
+  def render_with_attachments
+    final_pdf = ::CombinePDF.parse(self.render)
+
+    attachments.each do |attachment|
+      next unless File.extname(attachment.file.path) == ".pdf"
+
+      final_pdf << ::CombinePDF.load(attachment.file.path)
+    end
+
+    final_pdf.to_pdf
+  end
+end
