--- lib/active_record_comma_type_cast.rb
@@ -1,12 +1,0 @@
-module ActiveRecordCommaTypeCast
-  def cast_value(value)
-    if value.is_a?(String)
-      super value.gsub(",", ".").gsub(/[^-0-9\.]/, "")
-    else
-      super value
-    end
-  end
-end
-
-ActiveRecord::Type::Decimal.send(:prepend, ::ActiveRecordCommaTypeCast)
-ActiveRecord::Type::Float.send(:prepend, ::ActiveRecordCommaTypeCast)
