+++ lib/dorsale/core_and_rails_ext.rb
@@ -1,0 +1,3 @@
+require "dorsale/file_loader"
+require "dorsale/active_record_comma_type_cast"
+require "dorsale/form_back_url"
