+++ lib/dorsale/engine.rb
@@ -34,7 +34,6 @@ if Rails.env.test? || Rails.env.development?
 end
 
 require "acts-as-taggable-on"
-require "active_record_comma_type_cast"
 
 module Dorsale
   class Engine < ::Rails::Engine
@@ -62,16 +61,14 @@ module Dorsale
       Rails.application.config.assets.precompile += %w( dorsale/avatar.png )
     end
 
+    initializer "simple_form" do
+      require "dorsale/simple_form"
+      require "dorsale/simple_form_bootstrap"
+    end
+
     Mime::Type.register "application/vnd.ms-excel", :xls
     Mime::Type.register "application/vnd.ms-excel", :xlsx
   end
 end
 
-require "dorsale/file_loader"
-require "dorsale/polymorphic_id"
-require "dorsale/simple_form"
-require "dorsale/simple_form_bootstrap"
-require "dorsale/model_i18n"
-require "dorsale/model_to_s"
-require "dorsale/alexandrie/prawn"
-require "dorsale/form_back_url"
+require_relative "core_and_rails_ext"
